# openapi.yaml
openapi: 3.1.0
info:
  title: Console External API
  version: 0.1.0
  description: >-
    Curated external API surface for third-party developers. Authenticate every request with
    `Authorization: Bearer hbr_…` (API keys are minted in the Console web app).
servers:
  - url: https://api.testnet.console.walrus.xyz
    description: Testnet
...
paths:
  /api/v1/spaces:
    get:
      tags:
        - Spaces
      summary: List spaces
      description: List all spaces accessible to the authenticated user (personal + team).
      operationId: listSpaces
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - personal
              - team
          description: Filter by space type
      responses:
        '200':
          description: List of spaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SpaceListItem'
        '401':
          description: Not authenticated
  /api/v1/spaces/{id}/buckets:
    get:
      tags:
        - Buckets
      summary: List buckets in a space
      operationId: listBucketsBySpace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Space UUID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of buckets to return.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque cursor returned by a previous page.
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 255
          description: Case-insensitive substring match on bucket name.
        - name: visibility
          in: query
          required: false
          schema:
            type: string
            enum:
              - public
              - private
          description: Filter buckets by visibility.
        - name: sortField
          in: query
          required: false
          schema:
            type: string
            enum:
              - name
              - date
              - storage
            default: date
          description: Field to sort by (date = createdAt).
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: Sort direction.
      responses:
        '200':
          description: Buckets in the space (paginated).
          content:
            application/json:
              schema:
                type: object
                required:
                  - buckets
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bucket'
                  next_cursor:
                    type:
                      - string
                      - 'null'
                    description: Opaque cursor for the next page; null when no more results.
        '400':
          description: Invalid UUID
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Space not found
    post:
      tags:
        - Buckets
      summary: Create bucket in a space
      operationId: createBucket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Space UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - scope
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                scope:
                  type: string
                  enum:
                    - private
      responses:
        '201':
          description: >-
            Private bucket reserved. Sign the returned `bytes` locally with the service key and POST
            the signature to POST /buckets/{id}/finalize to complete creation.
          content:
            application/json:
              schema:
                type: object
                description: >-
                  Private bucket — reserve response. Sign `bytes` locally with the service key, then
                  POST the signature to /buckets/{id}/finalize.
                required:
                  - bucket_id
                  - bytes
                  - digest
                  - state
                properties:
                  bucket_id:
                    type: string
                    format: uuid
                  bytes:
                    type: string
                    description: Base64 sponsored transaction bytes.
                  digest:
                    type: string
                  state:
                    type: string
                    enum:
                      - pending_policy
        '400':
          description: Validation error
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Space not found
        '409':
          description: Bucket name conflict
        '422':
          description: Plan limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanLimitExceededResponse'
...
  /api/v1/buckets/{id}:
    get:
      tags:
        - Buckets
      summary: Get bucket by ID
      operationId: getBucketById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
      responses:
        '200':
          description: Bucket details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Bucket'
        '400':
          description: Invalid UUID
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Bucket not found
    put:
      tags:
        - Buckets
      summary: Update bucket
      operationId: updateBucket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                sealPolicyId:
                  type:
                    - string
                    - 'null'
                  minLength: 1
                  maxLength: 255
                visibility:
                  type: string
                  enum:
                    - public
                    - private
      responses:
        '200':
          description: Bucket updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UpdateBucketResponse'
        '400':
          description: Validation error
        '401':
          description: Not authenticated
        '403':
          description: Access denied or visibility change attempted (immutable in v1)
        '404':
          description: Bucket not found
        '409':
          description: Bucket name conflict
    delete:
      tags:
        - Buckets
      summary: Delete bucket
      description: Requires query parameter `confirm=true`.
      operationId: deleteBucket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
      responses:
        '204':
          description: Bucket deleted
          headers:
            X-Deleted-Id:
              description: UUID of the deleted bucket
              schema:
                type: string
                format: uuid
        '400':
          description: >-
            Missing confirm=true, validation error, or bucket is not empty (files still present or
            storage in use).
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Bucket not found
  /api/v1/buckets/{id}/files:
    get:
      tags:
        - Files
      summary: List files in bucket
      operationId: listBucketFiles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque cursor from a previous response's next_cursor
        - name: q
          in: query
          schema:
            type: string
            maxLength: 255
        - name: sortField
          in: query
          schema:
            type: string
            enum:
              - name
              - date
              - size
              - type
            default: date
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                      hasMore:
                        type: boolean
                      nextCursor:
                        type:
                          - string
                          - 'null'
                        description: Opaque cursor for the next page, null when no more results
        '400':
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Files
      summary: Upload file (async)
      operationId: uploadBucketFile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                  maxLength: 255
                metadata:
                  type: string
                  description: Optional JSON object string persisted in Console metadata column. Maximum 8KB.
                  example: '{"tag":"blue"}'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/FileSummary'
                required:
                  - data
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Bucket not found
        '409':
          description: Bucket not finalized or duplicate file name in bucket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited
  /api/v1/buckets/{id}/finalize:
    post:
      tags:
        - Buckets
      summary: Finalize private bucket
      description: >-
        Submit the on-chain signature to finalize a pending private bucket. Transitions the bucket
        from 'pending' to 'active'.
      operationId: finalizeBucket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - signature
              properties:
                signature:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Bucket finalized
          content:
            application/json:
              schema:
                type: object
                properties:
                  bucket_id:
                    type: string
                    format: uuid
                  seal_policy_id:
                    type:
                      - string
                      - 'null'
                  state:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Bucket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/buckets/{id}/files/{fileId}:
    get:
      tags:
        - Files
      summary: Get file metadata
      operationId: getFileById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: File UUID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/FileSummary'
                required:
                  - data
        '400':
          description: Invalid id
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found (or file not in this bucket)
    delete:
      tags:
        - Files
      summary: Delete file (async, soft-delete + worker)
      description: >-
        Marks the file as `deleting` and enqueues a background job that asks Oyster to remove the
        underlying blob. Returns 204 immediately; storage is released only after the worker confirms
        the Oyster delete. Subsequent retries are deduplicated by file id.
      operationId: deleteBucketFile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Bucket UUID
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: File UUID
      responses:
        '204':
          description: Accepted; file marked deleting
        '400':
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: File not found in this bucket
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key passed as Bearer token. Created via POST /api/v1/api-keys.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code (when available)
          enum:
            - USED_NONCE
            - EXPIRED_TIMESTAMP
            - ADDRESS_MISMATCH
            - INVALID_CHALLENGE
            - unauthorized
            - api_key_registering
            - api_key_revoking
            - api_key_revoked
            - read_only_api_key
            - bucket_not_in_scope
            - mirror_missing_grant
            - bucket_not_finalized
            - quota_exceeded
            - payload_too_large
            - bad_request
      required:
        - error
    PlanLimitExceededResponse:
      type: object
      required:
        - code
        - limit
        - currentTier
      properties:
        code:
          type: string
          enum:
            - PLAN_LIMIT_EXCEEDED
        limit:
          type: string
          enum:
            - storage
            - users
            - buckets
        currentTier:
          type: string
          enum:
            - free
...
    SpaceListItem:
      type: object
      required:
        - id
        - type
        - name
        - storage_used
        - storage_cap
        - bucket_count
        - role
        - created_at
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - personal
            - team
        name:
          type: string
        storage_used:
          type: integer
          description: Bytes currently stored
        storage_cap:
          type: integer
          description: Maximum bytes allowed
        bucket_count:
          type: integer
        member_count:
          type:
            - integer
            - 'null'
          description: Only present for team spaces
        role:
          type: string
          enum:
            - owner
            - admin
            - editor
            - viewer
          description: owner for personal, team role for team spaces
        created_at:
          type: string
          format: date-time
    Bucket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        space_id:
          type: string
          format: uuid
        name:
          type: string
        oyster_bucket_name:
          type: string
        visibility:
          type: string
          enum:
            - public
            - private
        seal_policy_id:
          type:
            - string
            - 'null'
        storage_used:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - space_id
        - name
        - oyster_bucket_name
        - visibility
        - seal_policy_id
        - storage_used
        - storage_cap
        - created_at
        - updated_at
...
    UpdateBucketResponse:
      type: object
      required:
        - id
        - name
        - visibility
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        visibility:
          type: string
          enum:
            - public
            - private
        updated_at:
          type: string
          format: date-time
    FileSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bucket_id:
          type: string
          format: uuid
        name:
          type: string
        blob_id:
          type:
            - string
            - 'null'
        oyster_object_id:
          type:
            - string
            - 'null'
        size:
          type: integer
        mime_type:
          type:
            - string
            - 'null'
        status:
          type: string
        is_private:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
          description: User-defined metadata persisted by Console.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type:
            - string
            - 'null'
          format: date-time
      required:
        - id
        - bucket_id
        - name
        - size
        - status
        - is_private
        - created_at
        - updated_at